+++ modules/features/discover/src/main/java/au/com/shiftyjelly/pocketcasts/discover/view/DiscoverAdapter.kt
@@ -121,6 +121,16 @@ internal class DiscoverAdapter(
         fun cancelLoading() {
             loadingDisposable?.dispose()
         }
+
+        open fun onRestoreInstanceState(state: Parcelable?) {
+            if (state != null) {
+                recyclerView?.layoutManager?.onRestoreInstanceState(state)
+            }
+        }
+
+        fun onSaveInstanceState(): Parcelable? {
+            return recyclerView?.layoutManager?.onSaveInstanceState()
+        }
     }
 
     interface ShowAllRow {
@@ -147,9 +157,12 @@ internal class DiscoverAdapter(
     inner class CarouselListViewHolder(var binding: RowCarouselListBinding) : NetworkLoadableViewHolder(binding.root) {
         val adapter = CarouselListRowAdapter(null, theme, listener::onPodcastClicked, listener::onPodcastSubscribe)
 
+        private val linearLayoutManager =
+            LinearLayoutManager(itemView.context, RecyclerView.HORIZONTAL, false).apply {
+                initialPrefetchItemCount = 1
+            }
+
         init {
-            val linearLayoutManager = LinearLayoutManager(itemView.context, RecyclerView.HORIZONTAL, false)
-            linearLayoutManager.initialPrefetchItemCount = 1
             recyclerView?.layoutManager = linearLayoutManager
             recyclerView?.itemAnimator = null
             val snapHelper = HorizontalPeekSnapHelper(0)
@@ -161,6 +174,13 @@ internal class DiscoverAdapter(
             recyclerView?.adapter = adapter
             adapter.submitList(listOf(LoadingItem()))
         }
+
+        override fun onRestoreInstanceState(state: Parcelable?) {
+            super.onRestoreInstanceState(state)
+            recyclerView?.post {
+                binding.pageIndicatorView.position = linearLayoutManager.findFirstVisibleItemPosition()
+            }
+        }
     }
 
     inner class SmallListViewHolder(val binding: RowPodcastSmallListBinding) : NetworkLoadableViewHolder(binding.root), ShowAllRow {
@@ -169,9 +189,12 @@ internal class DiscoverAdapter(
         override val showAllButton: TextView
             get() = binding.btnShowAll
 
+        private val linearLayoutManager =
+            LinearLayoutManager(itemView.context, RecyclerView.HORIZONTAL, false).apply {
+                initialPrefetchItemCount = 2
+            }
+
         init {
-            val linearLayoutManager = LinearLayoutManager(itemView.context, RecyclerView.HORIZONTAL, false)
-            linearLayoutManager.initialPrefetchItemCount = 2
             recyclerView?.layoutManager = linearLayoutManager
             recyclerView?.itemAnimator = null
             val snapHelper = HorizontalPeekSnapHelper(0.dpToPx(itemView.context))
@@ -183,6 +206,13 @@ internal class DiscoverAdapter(
             recyclerView?.adapter = adapter
 
             adapter.showLoadingList()
+        }
+
+        override fun onRestoreInstanceState(state: Parcelable?) {
+            super.onRestoreInstanceState(state)
+            recyclerView?.post {
+                binding.pageIndicatorView.position = linearLayoutManager.findFirstVisibleItemPosition()
+            }
         }
     }
 
@@ -267,7 +297,7 @@ internal class DiscoverAdapter(
                         loadPodcastList(row.source),
                         onNext = {
                             holder.adapter.fromListId = row.listUuid
-                            holder.adapter.submitList(it.podcasts)
+                            holder.adapter.submitList(it.podcasts) { onRestoreInstanceState(holder) }
                         }
                     )
                     row.listUuid?.let { AnalyticsHelper.listImpression(it) }
@@ -294,7 +324,7 @@ internal class DiscoverAdapter(
                         loadingFlowable,
                         onNext = {
                             holder.adapter.pillText = row.title.tryToLocalise(resources)
-                            holder.adapter.submitList(it)
+                            holder.adapter.submitList(it) { onRestoreInstanceState(holder) }
                             holder.binding.pageIndicatorView.count = it.count()
                         }
                     )
@@ -309,7 +339,7 @@ internal class DiscoverAdapter(
                             val podcasts = it.podcasts.subList(0, Math.min(MAX_ROWS_SMALL_LIST, it.podcasts.count()))
                             holder.binding.pageIndicatorView.count = Math.ceil(podcasts.count().toDouble() / SmallListRowAdapter.SmallListViewHolder.NUMBER_OF_ROWS_PER_PAGE.toDouble()).toInt()
                             holder.adapter.fromListId = row.listUuid
-                            holder.adapter.submitPodcastList(podcasts)
+                            holder.adapter.submitPodcastList(podcasts) { onRestoreInstanceState(holder) }
                         }
                     )
                     row.listUuid?.let { AnalyticsHelper.listImpression(it) }
@@ -321,7 +351,7 @@ internal class DiscoverAdapter(
                     holder.loadSingle(
                         service.getCategoriesList(row.source),
                         onSuccess = {
-                            adapter.submitList(it)
+                            adapter.submitList(it) { onRestoreInstanceState(holder) }
                         }
                     )
                 }
@@ -364,6 +394,7 @@ internal class DiscoverAdapter(
 
                             val textSize = if ((podcastTitle ?: "").length < 15) 18f else 15f
                             holder.binding.lblTitle.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSize)
+                            onRestoreInstanceState(holder)
                         }
                     )
                 }
@@ -423,6 +454,7 @@ internal class DiscoverAdapter(
                                 }
                                 listener.onEpisodeClicked(episode = episode, listUuid = row.listUuid)
                             }
+                            onRestoreInstanceState(holder)
                             row.listUuid?.let { listUuid -> AnalyticsHelper.listImpression(listUuid) }
                         }
                     )
@@ -469,6 +501,8 @@ internal class DiscoverAdapter(
 
                             holder.binding.lblSubtitle.text = it.subtitle?.tryToLocalise(resources)?.uppercase(Locale.getDefault())
 
+                            onRestoreInstanceState(holder)
+
                             row.listUuid?.let { listUuid -> AnalyticsHelper.listImpression(listUuid) }
                         }
                     )
@@ -477,12 +511,6 @@ internal class DiscoverAdapter(
 
             if (holder is ShowAllRow) {
                 holder.showAllButton.setOnClickListener { listener.onPodcastListClicked(row) }
-            }
-
-            if (holder is NetworkLoadableViewHolder) {
-                savedState[holder.itemId]?.let {
-                    holder.recyclerView?.layoutManager?.onRestoreInstanceState(it)
-                }
             }
         } else if (row is ChangeRegionRow) {
             val changeRegionRowViewHolder = holder as ChangeRegionViewHolder
@@ -498,12 +526,16 @@ internal class DiscoverAdapter(
     }
 
     private val savedState: MutableMap<Long, Parcelable?> = mutableMapOf()
+
+    private fun onRestoreInstanceState(holder: NetworkLoadableViewHolder) {
+        holder.onRestoreInstanceState(savedState[holder.itemId])
+    }
     override fun onViewRecycled(holder: RecyclerView.ViewHolder) {
         super.onViewRecycled(holder)
         when (holder) {
             is NetworkLoadableViewHolder -> {
                 holder.cancelLoading()
-                savedState[holder.itemId] = holder.recyclerView?.layoutManager?.onSaveInstanceState()
+                savedState[holder.itemId] = holder.onSaveInstanceState()
             }
         }
     }