+++ RELEASE-NOTES.md
@@ -56,6 +56,8 @@
         ([#77](https://github.com/Automattic/pocket-casts-android/pull/77)).
     *   Fix showing paused downloads as in progress
         ([#113](https://github.com/Automattic/pocket-casts-android/pull/113)).
+    *   Fix issues with discover feed state when scrolling through the feed
+        ([#120](https://github.com/Automattic/pocket-casts-android/pull/120)).
 
 ### 7.19.2 (2022-02-11)